Rename ban-api into ban-api-depot
